qemu: Don't wait for the end of the VM
This patch brings in a real improvement about the time it takes
to delete the pod. This is because it does not wait anymore for
the end of the VM. Indeed, as long as we have called into QMP to
shutdown the VM, this is not Clear Containers responsibility if
the VM cannot be stopped for any reason. It basically should not
happen.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>